remove BIP70 APIs remove generateInvoice from wallet interface
