+++ src/main/scala/registry.scala
@@ -59,8 +59,6 @@ package object registry {
 
     Await.result(setupDb, 10 seconds)
 
-    //    def offerTable = OfferRepositoryRegistry.offerRepositoryImpl.offersTable
-    //    def sessionTable = SessionRepositoryRegistry.sessionRepositoryImpl.sessionsTable
 
     def setupDb = database.db.run({
       logger.info(s"Setting up schemas and populating tables")
